sparql endpoint
Created a new connection between graphDB (the sparql endpoint) and python. The question (triple) is saved to hints in case of a 'yes' answer. Each time we make a subGraph we utilize the triples saved in hints to make a query that get all triples which have the same predicate object combination as the triples in hints.
TODO: exclude the predicate object combination from the previously asked questions (so we dont ask the same questions twice).